+++ primefaces/src/main/resources/META-INF/resources/primefaces/tabview/tabview.js
@@ -60,6 +60,7 @@
  * @prop {number} cfg.tabindex Position of the element in the tabbing order.
  * @prop {boolean} cfg.multiViewState Whether to keep TabView state across views.
  * @prop {boolean} cfg.focusOnError Whether to focus the first tab that has an error associated to it.
+ * @prop {boolean} cfg.focusOnLastActiveTab Whether to focus on the last active tab that a user selected.
  */
 PrimeFaces.widget.TabView = PrimeFaces.widget.DeferredWidget.extend({
 
@@ -336,13 +337,15 @@ PrimeFaces.widget.TabView = PrimeFaces.widget.DeferredWidget.extend({
     },
 
     /**
-     * Binds refresh listener to update error highlighting on component udpate.
+     * Binds refresh listener to update error highlighting or restore the last active tab on component udpate.
      * @private
      */
     bindRefreshListener: function() {
         var $this = this;
         var focusIndex = -1;
         this.addRefreshListener(function() {
+			
+			// update error highlighting and set focusIndex
             $(this.jqId + '>ul>li').each(function() {
                 var tabId = $('a', this).attr('href').slice(1);
                 tabId = PrimeFaces.escapeClientId(tabId);
@@ -355,7 +358,14 @@ PrimeFaces.widget.TabView = PrimeFaces.widget.DeferredWidget.extend({
                     $(this).removeClass('ui-state-error');
                 }
             });
-            if ($this.cfg.focusOnError && focusIndex >= 0) {
+            
+            // set focusIndex to restore the last active tab
+            // "focusOnError" always takes precedence over "focusOnLastActiveTab"
+            if (focusIndex < 0) {
+				focusIndex = $this.cfg.selected;
+			}
+
+            if (($this.cfg.focusOnError || $this.cfg.focusOnLastActiveTab) && focusIndex >= 0) {
                setTimeout(function () {$this.select(focusIndex, true)}, 10);
             }
         });